# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/business/status": {
            "get": {
                "description": "사업자번호로 매장의 상태를 조회한다.",
                "tags": [
                    "Business API"
                ],
                "summary": "휴폐업 상태 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "사업자번호",
                        "name": "biz_number",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.CodefData"
                        }
                    }
                }
            }
        },
        "/deposit": {
            "get": {
                "description": "월단위 기준으로 일자별 입금금액을 조회한다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Deposit API"
                ],
                "summary": "일자별 입금금액 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "연도 - ex)2021",
                        "name": "year",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "월 - ex)04",
                        "name": "month",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/deposit/:date": {
            "get": {
                "description": "일자 기준으로 카드사별 입금금액을 조회한다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Deposit API"
                ],
                "summary": "카드사별 입금금액 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "일자",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.CodefData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "CF-00000"
                },
                "extraMessage": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "성공"
                },
                "resBusinessStatus": {
                    "type": "string",
                    "example": "부가가치세 일반과세자 입니다."
                },
                "resClosingDate": {
                    "type": "string"
                },
                "resCompanyIdentityNo": {
                    "type": "string",
                    "example": "1231212345"
                },
                "resTaxationTypeCode": {
                    "type": "string",
                    "example": "1"
                },
                "resTransferTaxTypeDate": {
                    "type": "string"
                }
            }
        }
    }
}